Drop the depth knob and the unused conductor surface.
One Artefato path: plenitude — cover the facets, none folded. report-deep and kin stay aliases of that path, not a second target.
