+++ skills/discovery-deep/SKILL.md
@@ -1,14 +1,8 @@
 ---
 name: discovery-deep
-description: The `discovery` producer at DEEP depth — a serendipitous find explored across multiple
-  applications and angles, not the tight single-point `brief` default. A discoverable alias for
-  /{prefix}-discovery at depth=deep. Invoked as /{prefix}-discovery-deep or run inside the beat.
+description: Discoverable alias for the discovery producer. Same genus path as /{prefix}-discovery —
+  plenitude, every facet of the find covered. Invoked as /{prefix}-discovery-deep or run inside the beat.
 ---
-You are the **discovery** producer run at the **`deep`** depth target (`scaffold.md`: Depth). This is a
-**thin alias, NOT a separate contract**: run the full `skills/discovery` skill exactly as written, with
-**one override**: the depth target is **`deep`**, not discovery's `brief` default.
-
-At `deep`: the ONE serendipitous insight is still the subject, but you **explore it across multiple
-applications** to the mentee's live work, work out the angles, trace where it does and doesn't transfer —
-the well-read friend who doesn't just hand you the thing but sits down and works through everything it
-could mean for you. The rich-rite floor is unchanged. Everything else follows `skills/discovery/SKILL.md`.
+You are the **discovery** producer. This is a **thin alias, NOT a separate contract** and NOT a
+depth override: run the full `skills/discovery` skill exactly as written. There is one production
+path — cover the facets, leave no facet folded. `/{prefix}-discovery-deep` names that same path.
