+++ skills/report-deep/SKILL.md
@@ -1,15 +1,8 @@
 ---
 name: report-deep
-description: The `report` producer at DEEP depth (plenitude) — a prose-synthesis Artefato developed to
-  its full depth. A discoverable alias for /{prefix}-report at depth=deep, for those who don't know depth
-  is a per-artefato override. Invoked as /{prefix}-report-deep or run inside the beat.
+description: Discoverable alias for the report producer. Same genus path as /{prefix}-report —
+  plenitude, every facet covered. Invoked as /{prefix}-report-deep or run inside the beat.
 ---
-You are the **report** producer run at the **`deep`** depth target (`scaffold.md`: Depth). This is a
-**thin alias, NOT a separate contract**: run the full `skills/report` skill exactly as written — same
-gather-grounding → converge/diverge → close → publish, same rich-rite floor, same everything — with **one
-override**: the depth target is **`deep`** (plenitude), not report's `standard` default.
-
-At `deep`: develop **every** cognitive move to its full depth, draw out the load-bearing implications,
-leave **no facet folded**. Synthesis-to-a-bite is a failure here. The rich-rite floor (the four moves
-present) is unchanged — depth only sets how far *above* it you develop. Everything else follows
-`skills/report/SKILL.md` verbatim.
+You are the **report** producer. This is a **thin alias, NOT a separate contract** and NOT a
+depth override: run the full `skills/report` skill exactly as written. There is one production
+path — cover the facets, leave no facet folded. `/{prefix}-report-deep` names that same path.
