+++ skills/research-deep/SKILL.md
@@ -1,14 +1,8 @@
 ---
 name: research-deep
-description: The `research` producer at DEEP depth (plenitude) — a directed deep-dive developed to its
-  full depth. A discoverable alias for /{prefix}-research (whose default is already `deep`), named
-  explicitly for those who don't know depth is a per-artefato knob. Invoked as /{prefix}-research-deep.
+description: Discoverable alias for the research producer. Same genus path as /{prefix}-research —
+  plenitude, every facet covered. Invoked as /{prefix}-research-deep or run inside the beat.
 ---
-You are the **research** producer run at the **`deep`** depth target (`scaffold.md`: Depth). This is a
-**thin alias, NOT a separate contract**: run the full `skills/research` skill exactly as written, with the
-depth target pinned to **`deep`** (plenitude).
-
-Note: research's **default is already `deep`**, so this alias is behaviourally the same as
-`/{prefix}-research` — it exists purely for **discoverability** (a browsing operator sees `research-deep`
-and knows the deep-dive goes to plenitude). Develop the named target exhaustively: every move, every
-facet, no fold. The rich-rite floor is unchanged. Everything else follows `skills/research/SKILL.md`.
+You are the **research** producer. This is a **thin alias, NOT a separate contract** and NOT a
+depth override: run the full `skills/research` skill exactly as written. There is one production
+path — cover the facets, leave no facet folded. `/{prefix}-research-deep` names that same path.
